hdlcpp: Change private to protected to facilitate overrides
Prior to this change, overriding read or write made no sense because the framing and
encoding functions were behind a private wall and would require one to re-implement all
those functions unfortunately making this library less useful. For example, for a high
speed bus of 200 Mbps with 8b/10b encoding at the hardware layer, having a processor
wait for acks kills throughput. In that instance we are interested in the framing,
encoding and reading functions. This change makes such a scenario possible.

Signed-off-by: Edward Kigwana <ekigwana@gmail.com>