buffer: allow erase to include tail
Only allowing `erase` up to, not including, the tail pointer means that
an entire buffer clearing can never really happen unless the tail is
actually past-the-end of the underlying memory area.

Allow erases to cover the entire span from head to tail inclusively.

On a side note, there doesn't seem to be a good reason to dissallow even
erase where `last` is past the tail pointer. So no extra sanitization
logic is added for such cases.