Protect transportWrite with writeMutex
Guard usage of the `transportWrite` and encoding frames into the
`writeBuffer` by adding a `writeFrameMutex` to be locked within
`writeFrame`.

The `write` function relies on repeated calls to `read` that need to be
performed by a separate thread from the user of hdlcpp. This is because
`write` will block until a `frameAck` or `frameNack` is received or a
timeout is hit.
The hdlcpp implementation, however, also performs writes on the
transport layer from within the `read` function itself. It does so in
the process of either sending a `frameAck` or `frameNack` upon receival
of a data frame.

From this it is aparent that any user of hdlcpp is basically forced
into a race condition that can potentially mangle frames on the wire.
The race can be hit by simply receiving data frames while trying to
transmit data frames as well.
Since the function which actually encodes data into the `writeBuffer`
and calls the `transportWrite` callback is `writeFrame`, adding a mutex
to specifically guard this function can prevent the race. This makes
sure that if the thread handling `read` were to attempt transmitting a
`frameAck` while another thread was attempting to transmit a data frame,
then these two frames don't get mixed together.

By introducing a new mutex rather than just moving the existing
`writeMutex`, the thread safety of the sequence numbering increments
within `write` is kept as-is.